c json-ast: expose typed AST
